requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.14
fastapi==0.110.0
frozenlist==1.4.1
googleapis-common-protos==1.62.0
grpcio==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
hyperdx-opentelemetry==0.1.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.3.0
iniconfig==2.0.0
limits==3.10.0
multidict==6.0.5
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-grpc==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-instrumentation-sqlite3==0.43b0
opentelemetry-instrumentation-tortoiseorm==0.43b0
opentelemetry-instrumentation-urllib==0.43b0
opentelemetry-instrumentation-urllib3==0.43b0
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==23.2
pluggy==1.4.0
protobuf==4.25.3
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyOpenSSL==24.1.0
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.1.1
slowapi==0.1.9
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
